# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Europass CV Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cvs/{id}/languages:
    post:
      operationId: createLanguage
      tags: [languages]
      summary: Add a language to a CV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageInput"
      responses:
        "201":
          description: Created language
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Language"
  /cvs/{id}/languages/{langId}:
    put:
      operationId: updateLanguage
      tags: [languages]
      summary: Update a language entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: langId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageInput"
      responses:
        "200":
          description: Updated language
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Language"
    delete:
      operationId: deleteLanguage
      tags: [languages]
      summary: Delete a language entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: langId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cvs/{id}/export/pdf:
    get:
      operationId: exportCvPdf
      tags: [export]
      summary: Export CV as PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cvs/{id}/export/word:
    get:
      operationId: exportCvWord
      tags: [export]
      summary: Export CV as Word document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Word file
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Language:
      type: object
      properties:
        id:
          type: integer
        cvId:
          type: integer
        languageName:
          type: string
        listening:
          type: ["string", "null"]
        reading:
          type: ["string", "null"]
        spokenInteraction:
          type: ["string", "null"]
        spokenProduction:
          type: ["string", "null"]
        writing:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - id
        - cvId
        - languageName
        - sortOrder
    LanguageInput:
      type: object
      properties:
        languageName:
          type: string
          minLength: 1
        listening:
          type: ["string", "null"]
        reading:
          type: ["string", "null"]
        spokenInteraction:
          type: ["string", "null"]
        spokenProduction:
          type: ["string", "null"]
        writing:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - languageName